ref: rename schulden to kontopfaendung
